requirements.txt
allure-pytest
requests
pytest
pyautogui